make default method post and move imports
